working through styling recipe cards on recipes route
