+++ public/stylesheets/recipes.css
@@ -1,110 +1,12 @@
-/* background layout */
-.body {
-  background-color: #e7e0d6;
-}
-
-/*navbar styles*/
-.navbar {
+.main-body{
   display: flex;
-  justify-content: space-between;
-  padding: 10px;
-  padding-top: 40px;
-  align-content: center;
-  height: 100px;
-  box-shadow: 0 2px 2px rgba(0, 0, 0, 0.25);
-}
-
-.navbar__logo {
-  margin-left: -20px;
-  margin-bottom: 35px;
-}
-
-.navbar__text {
-  font-size: 24px;
-  margin: 10px;
-}
-
-.navbar__link {
-  font-size: 20px;
-}
-
-.navbar__container-right .navbar__link {
-  margin: 5px;
-  border: 1px solid black;
-  text-decoration: none;
-  padding: 4px;
-  height: 30px;
-  border-radius: 5px;
-}
-
-.navbar__container {
-  display: flex;
-  margin: 15px;
-  padding: 10px;
-  align-items: center;
-}
-
-.navbar__container-left {
-  display: flex;
-  justify-content: start;
-}
-
-.navbar__container-right {
-  display: flex;
-  justify-content: end;
-}
-
-.searchForm {
-  display: inline;
-  box-sizing: border-box;
-}
-
-.searchbar {
-  border-radius: 5px;
-}
-
-#searchBar {
-  box-sizing: border-box;
-  height: 28px;
-}
-
-#searchIcon {
-  box-sizing: border-box;
-  height: 28px;
-  box-shadow: 1px 1px 2px 1px rgba(10, 128, 238, 0.25);
-  background-color: #ec4c4c;
-  font-weight: bold;
-}
-
-/*main section styling*/
-.main-body {
-  display: grid;
-  grid-template-columns: 1fr 2fr 1fr;
-  grid-template-rows: auto;
-  grid-template-areas: 'userCol recipeCol newsCol';
-}
-
-.main-body__main-column {
-  display: flex;
-  flex-direction: column;
-  margin: 10px 15px 1px 15px;
-  padding: 10px;
-  align-items: center;
-  box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-}
-
-.main-body__user-column {
-  grid-area: 'userCol';
-  padding-left: 20px;
-}
-
-.main-body__recipe-column {
-  grid-area: 'recipeCol';
-  display: flex;
-  justify-content: center;
-}
-.recipe {
-  text-align: center;
+  flex-direction: row;
+  flex-wrap: wrap;
+  width: 70%;
+  margin: auto;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  padding: 15px;
 }
 
 .recipe-card {
@@ -120,99 +22,37 @@
 
 .recipe-card {
   display: grid;
-  height: 450px;
-  width: 800px;
-  grid-template-columns: 2fr 1fr;
+  height: 270px;
+  width: 30%;
+  grid-template-columns: 2fr 3fr;
   grid-template-rows: auto;
   grid-template-areas:
-    'recipeImg recipeTxt'
     'recipeImg recipeTxt'
     'recipeImg recipeTxt'
     'recipeImg recipeTxt'
     'recipeImg recipeSave';
 }
 
-.recipe-image {
+.image-link__container {
   grid-area: recipeImg;
-  margin: 20px;
-  align-items: center;
+  margin: 10px;
+  text-align: left;
 }
+
+.recipe-image > img{
+  width: 100%;
+  height: auto;
+}
+
 .recipe-content {
   grid-area: recipeTxt;
   margin: 10px;
+  padding: 10px;
+  overflow: hidden;
 }
 .recipe-description {
   text-align: left;
 }
 .recipe-save {
   grid-area: recipeSave;
-}
-
-.main-body__news-column {
-  grid-area: newsCol;
-  padding-right: 20px;
-}
-
-.main-body__user-column > .recipe-image {
-  height: auto;
-  width: 75%;
-}
-
-.main-body__recipe-column .recipe-image {
-  height: auto;
-  width: 50%;
-}
-
-.footer {
-  background-color: #3c6382;
-  display: flex;
-  width: 100%;
-  /* text-transform: capitalize; */
-  padding-top: 25px;
-  padding-bottom: 25px;
-  margin-top: 80px;
-}
-
-
-.rating {
-  unicode-bidi: bidi-override;
-  direction: rtl;
-  padding: 10px;
-}
-.rating > span {
-  display: inline-block;
-  position: relative;
-  width: 1.1em;
-}
-.rating > span:hover:before,
-.rating > span:hover ~ span:before {
-   content: "\2605";
-   position: absolute;
-}
-.footer__div {
-  width: 25%;
-  padding: 20px;
-
-}
-.footer__h3 {
-  color:black;
-  line-height: 1.5;
-  letter-spacing: 1px;
-  font-weight: bold;
-}
-.footer__link {
-  text-decoration: none;
-  color: black;
-  padding: 5px;
-  line-height: 2;
-}
-.footer__link:hover {
-  color: #ec4c4c;
-}
-li {
-  list-style-type: none;
-}
-.footer__p {
-  line-height: 1.5;
-  letter-spacing: 1.5px;
 }
