+++ views/layout.pug
@@ -6,7 +6,6 @@ html
       link(rel='stylesheet', href='/stylesheets/home.css')
       link(rel='stylesheet', href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.11.1/css/all.css')
       link(rel='stylesheet', href='/stylesheets/reset.css')
-      link(rel='stylesheet', href='/stylesheets/recipe.css')
   header(class='header')
       nav(class='navbar core-display')
         div(class='navbar__container navbar__container-left')
