+++ views/utils.pug
@@ -19,9 +19,10 @@ mixin saveButtons(recipe)
       button(class='recipe-favorited button', id=`favoriterecipe:${recipe.id}.cupboard:${user.Cupboards[0].id}`, type='submit') Favorite
 
 mixin ratingButtons(recipe)
-  div(class='rating')
-    span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:5`)
-    span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:4`)
-    span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:3`)
-    span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:2`)
-    span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:1`)
+  if locals.authenticated
+    div(class='rating')
+      span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:5`)
+      span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:4`)
+      span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:3`)
+      span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:2`)
+      span(class='rating__button fa fa-star', id=`recipe:${recipe.id}.star:1`)
